rename sign up page and organize code
